file-sharing-app-backend/requirements.txt
asgiref==3.8.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
colorama==0.4.6
cryptography==44.0.0
Django==5.1.4
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
idna==3.10
pillow==11.0.0
pycparser==2.22
PyJWT==2.10.1
pyotp==2.9.0
python-dotenv==1.0.1
qrcode==8.0
requests==2.32.3
sqlparse==0.5.2
tzdata==2024.2
urllib3==2.2.3
